+++ tcs_garr/commands/list_certificates.py
@@ -9,7 +9,7 @@ from dateutil import parser
 from tabulate import tabulate
 
 from tcs_garr.commands.base import BaseCommand
-from tcs_garr.utils import CertificateStatus, load_config
+from tcs_garr.utils import CertificateStatus, UserRole, load_config
 
 
 class ListCertificatesCommand(BaseCommand):
@@ -21,6 +21,8 @@ class ListCertificatesCommand(BaseCommand):
     expired since a certain number of days or certificates expiring in the
     next few days.
     """
+
+    REQUIRED_ROLE = UserRole.USER
 
     def __init__(self, args):
         """
@@ -75,7 +77,18 @@ class ListCertificatesCommand(BaseCommand):
             nargs="?",
             const=True,
             default=None,
-            help="Filter certificates owner by user. Without arg (--user only) will filter for the logged in user.",
+            help=(
+                "Filter certificates owner by user. Without arg (--user only) will "
+                "filter for the logged in user. "
+                "Use this if you have Approver role or Admin role."
+            ),
+        )
+
+        # Add an optional argument to filter certificates by FQDN
+        parser.add_argument(
+            "--fqdn",
+            type=str,
+            help="Filter certificates by a substring in their Fully Qualified Domain Name (FQDN).",
         )
 
         # Add export flag
@@ -84,6 +97,42 @@ class ListCertificatesCommand(BaseCommand):
             action="store_true",
             help="Export certificates to json file.",
         )
+
+    def _filter_certificates(self, certificates, statuses, username=None):
+        """
+        Apply filters to the list of certificates based on status, username, and FQDN.
+
+        Args:
+            certificates (list): List of certificate dictionaries.
+            statuses (list): List of statuses to filter certificates by.
+            username (str, optional): Username to filter certificates by.
+
+        Returns:
+            list: Filtered list of certificates.
+            dict: Recap of filtered certificates.
+        """
+        # Filter by username
+        if username:
+            certificates = [cert for cert in certificates if cert["userEmail"] == username]
+
+        # Filter by status
+        certificates = [cert for cert in certificates if cert.get("status", "") in [status.value for status in statuses]]
+
+        # Filter by FQDN
+        fqdn_filter = self.args.fqdn
+        if fqdn_filter:
+            certificates = [
+                cert for cert in certificates if any(fqdn_filter in domain["fqdn"] for domain in cert.get("domains", []))
+            ]
+
+        # Build recap
+        recap = {"count": len(certificates)}
+        for cert in certificates:
+            status_name = cert["status"]
+            recap.setdefault(status_name, 0)
+            recap[status_name] += 1
+
+        return certificates, recap
 
     def get_cn_value(self, item):
         """
@@ -131,6 +180,52 @@ class ListCertificatesCommand(BaseCommand):
         # Return cn_value or "CN not found"
         return cn_value if cn_value else "CN not found"
 
+    def list_certificates_as_admin(self, client, username, statuses):
+        """
+        List certificates as admin user
+
+        Args:
+            client (HaricaClient): A client instance
+            username (str): Username to filter certificates by
+            statuses (list): List of certificate statuses to filter
+
+        Returns:
+            list: List of certificates
+            dict: Recap
+        """
+        certificates = []
+
+        for status in statuses:
+            start_index = 0
+            while True:
+                response = client.list_certificates(
+                    start_index=start_index,
+                    status=status,
+                )
+
+                if not response:
+                    break
+
+                certificates.extend(response)
+                start_index += len(response)
+
+        return self._filter_certificates(certificates, statuses, username)
+
+    def list_certificates_as_user(self, client, statuses):
+        """
+        List certificates as a regular user.
+
+        Args:
+            client (HaricaClient): A client instance.
+            statuses (list): List of certificate statuses to filter.
+
+        Returns:
+            list: List of certificates.
+            dict: Recap.
+        """
+        certificates = client.list_user_certificates()
+        return self._filter_certificates(certificates, statuses)
+
     def execute(self):
         """
         Execute the list certificates command to retrieve and display certificates.
@@ -154,7 +249,7 @@ class ListCertificatesCommand(BaseCommand):
         to_date = current_date + timedelta(days=self.args.expiring_in) if self.args.expiring_in is not None else None
 
         # Get the Harica client instance
-        harica_client = self.harica_client()
+        client = self.harica_client()
 
         # Get username if specified in args
         # True when --user without arg
@@ -173,32 +268,18 @@ class ListCertificatesCommand(BaseCommand):
         )
 
         # Retrieve the list of certificates from the Harica client
-        # Handle pagination and statues
-        certificates = []
-        recap = {
-            "count": 0,
-        }
+        # Handle pagination and statues if admin
 
-        for status in statuses:
-            start_index = 0
-            recap[status.name] = 0
-            while True:
-                response = harica_client.list_certificates(
-                    start_index=start_index,
-                    status=status,
-                )
-
-                if not response:
-                    break
-
-                certificates.extend(response)
-                start_index += len(response)
-                recap["count"] += len(response)
-                recap[status.name] += len(response)
-
-        if username:
-            # Filter certificates by username and the item field is userEmail
-            certificates = [cert for cert in certificates if cert["userEmail"] == username]
+        # if user role is only USER and does not have any other role
+        # get only user certificates
+        if client.has_role(UserRole.USER) and len(client.roles) == 1:
+            certificates, recap = self.list_certificates_as_user(client, statuses)
+        else:
+            certificates, recap = self.list_certificates_as_admin(
+                client,
+                username,
+                statuses,
+            )
 
         # Replace None value with datemin and convert to string
         # This will facilitate sorting
@@ -226,19 +307,15 @@ class ListCertificatesCommand(BaseCommand):
                 # Determine the Common Name value
                 cn_value = self.get_cn_value(item)
 
-                # Append the certificate data to the list for tabular display
-                data.append(
-                    [
-                        item["transactionId"],
-                        cn_value,
-                        item["certificateValidTo"],  # Expiration date of the certificate
-                        item["status"],
-                        (
-                            ";\n".join(subjAltName["fqdn"] for subjAltName in item["domains"]) if "domains" in item else ""
-                        ),  # Alternate names
-                        item["user"],  # User who requested the certificate
-                    ]
-                )
+                transaction_id = item["transactionId"]
+                expiration_date = item["certificateValidTo"]
+                status = item["status"]
+                status_info = item["transactionStatusMessage"]
+                user = item["user"]
+
+                alt_names = ";\n".join(domain["fqdn"] for domain in item.get("domains", []))
+
+                data.append([transaction_id, cn_value, expiration_date, status, status_info, alt_names, user])
 
         # Log the results in a formatted table with headers, using color for column titles
         self.logger.info(
@@ -249,16 +326,17 @@ class ListCertificatesCommand(BaseCommand):
                     "Common Name",
                     "Expire at",
                     "Status",
+                    "Info",
                     "Alt Names",
                     "Requested by" + Style.RESET_ALL,
                 ],
                 tablefmt="grid",
-                maxcolwidths=[None, 32, None, None, None, 20] if data else None,
+                maxcolwidths=[None, 32, None, None, 32, None, 20] if data else None,
             )
         )
 
         # Write a recap of certificate count by status
         self.logger.info(f"Total certificates: {recap['count']}")
         for status in CertificateStatus:
-            if status.name in recap:
-                self.logger.info(f"Certificates with status {status.name}: {recap[status.name]}")
+            if status.value in recap:
+                self.logger.info(f"Certificates with status {status.name}: {recap[status.value]}")